AST+Parser+Symbol+Syntax: Refactor SimpleIdentifier to guarantee a Symbol
The previous iteration of this was just a wrapper around a single Token value.

This presented problems when working with SimpleIdentifier values as you needed
to consistently `unwrap()` the symbol as it's optional on the Token. This new
API will destructure the Token into a Symbol, IdentifierQualification and Span
so that those nasty code smell `unwrap()` calls are no longer required.