Symbol: Add a new singleton API for SymbolTable
The problem with using symbols is that they are annoying to debug or display,
since they're just integers pointing to magical strings.

With this change, we can use a new SymbolTable::the() method to grab a singleton
instance of a SymbolTable using some "unsafe" Rust. This will allow us to impl
Debug and Display for the new struct based Symbol and provide more accurate output.