Various: Unimplement SymbolTable + Symbol
The SymbolTable was a hopeful way of reducing memory usage during
various phases of the PXP lifecycle. It worked to improve memory
usage, as well as parse speed, but the ergonomic effects that it
had aren't worth it.

I'm more interested in making the AST easy to use, especially in
threaded contexts. My next goal is to find a nice way of improving
allocation times, probably using an "arena allocator" or "bump" allocator.