Parser: Resolve an issue with binary operators and right-hand assignments being parsed with wrong precedence
Closes #73.

This one is a bit of a hack. I did some digging inside of microsoft/tolerant-php-parser
and found a similar thing going on there, so I'm giving it a go here to hopefully fix
this problem.

In an ideal world, this will be a one-and-done solution but I fear there may be some
edge cases to cover, such as certain left-hand expressions not needing this treatment.