Lexer: Improve nowdoc tokenisation sequence
We had this rather ugly `set_peek()` method that would set the peek
token so it was already there the next time you called `next()`.

This was causing issues as it was being placed into `peek` but then
peek was immediately being set to `None`, losing that piece of info.